code/requirements.txt
python-pptx
deepl
tkinter